requirements.txt
google-genai
sentence-transformers
numpy
python-dotenv
pydantic
pytest
